+++ packages/realtime-compiler/src/Routing/PageRouter.php
@@ -82,12 +82,10 @@ class PageRouter
     protected function getPageFromRoute(): HydePage
     {
         try {
-            return Routes::getOrFail($this->normalizePath($this->request->path))->getPage();
+            return Routes::get($this->normalizePath($this->request->path))->getPage();
         } catch (RouteNotFoundException $exception) {
-            $index = Routes::get($this->normalizePath($this->request->path).'/index');
-
-            if ($index) {
-                return $index->getPage();
+            if (Routes::exists($this->normalizePath($this->request->path).'/index')) {
+                return Routes::get($this->normalizePath($this->request->path).'/index')->getPage();
             }
 
             throw $exception;